1) Introduced a source file for maintaining utilites (xutils.cpp) 2) Added test for extract_filename function
